Tests/LibWeb/Text/expected/scrollend-after-keyboard-scroll.txt

scroll events received: true
scrollend fired at document: true
scrollend count: 1
window scrolled: true
smooth keyboard scroll moved: true
smooth keyboard scrollend: received
